Spider-master/requirements.txt
pre-commit==2.10.1
PyYAML==5.3.1
requests==2.25.1
